Unify get_top_of_ram and get_cbmem_toc

+++ src/northbridge/intel/gm45/ram_calc.c
@@ -83,7 +83,7 @@ u32 decode_igd_gtt_size(const u32 gsm)
 	}
 }
 
-u32 get_top_of_ram(void)
+unsigned long get_top_of_ram(void)
 {
 	const pci_devfn_t dev = PCI_DEV(0, 0, 0);
